+++ aeron-driver/src/main/c/aeronmd.h
@@ -297,7 +297,7 @@ typedef struct aeron_counters_manager_stct aeron_counters_manager_t;
 
 typedef int (*aeron_congestion_control_strategy_supplier_func_t)(
     aeron_congestion_control_strategy_t **strategy,
-    int32_t channel_length,
+    size_t channel_length,
     const char *channel,
     int32_t stream_id,
     int32_t session_id,
